refactor(core): extract SignatureEngine + verifyState to core
Move the universal signer (SignatureEngine class, ~330 lines) and the
audit-trail integrity verifier (verifyState, ~50 lines plus its private
helpers readAllKeyRecords, readAllDelegations, verifyDelegation,
verifyAuditRecord) from skills/alien-agent-id/lib.mjs into
plugins/agent-id-core/lib/signature-engine.mjs. The class's private
helpers (summarizePayload, safeName, agentKeyFile, delegationFile) and
the thin resolveAgentId context resolver move along with it.

SignatureEngine is the universal signer for any v3 attestation —
appendOperation builds the canonical envelope, signs it with the
agent's Ed25519 key, and writes a hash-chained audit-log entry.
Every plugin that produces a signed operation (git, vault, auth)
goes through this class.

verifyState is symmetric on the read side: walk the audit log and
report every chain-integrity, signature, or delegation failure.

lib.mjs re-exports the module with `export *`. No public-surface
changes; all 184 tests pass.
